+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 - [sdk/python] Fix bug with class methods for YAML transformations (https://github.com/pulumi/pulumi-kubernetes/pull/2469)
 - Fix StatefulSet await logic for OnDelete update (https://github.com/pulumi/pulumi-kubernetes/pull/2473)
+- Skip wait for Pods on headless Service (https://github.com/pulumi/pulumi-kubernetes/pull/2475)
 
 ## 3.29.1 (June 14, 2023)